requirements.txt
multiprocessing
RPi.GPIO
requests
